Rework object store to use runtime Store
